apps/savanaflow/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
reportlab==4.1.0
aiosmtplib==3.0.1
email-validator==2.1.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.4.4
python-dateutil==2.9.0

# Async task queue
celery[redis]==5.4.0
redis>=4.5.2,<5.0.0
